requirements.txt
Django==3.2
setuptools
fontawesomefree==6
psycopg2
django-jsonfield
django-extensions
boto3
factory-boy
pre-commit
flake8
django-html
Beautify
python-decouple
stripe
django-environ
django-filter
sendgrid